breaks down Home into multiple components
